Feature: ReadMe's complete
1. V1 of Readmes complete with the following:
An explanation of the what the app does and how it works with the following headers:
A link to the other repo
A link to both deployed sites
List of technologies used
List unsolved problems which would be fixed in future iterations.
Document your planning, process and problem-solving strategy
Link to Entity Relationship Diagram (ERD).
Set up and installation instructions for back end application